requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
APScheduler==3.9.1
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
backports.zoneinfo==0.2.1
baidu-aip==2.2.18.0
betterproto==1.2.5
blinker==1.4
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.0.4
distlib==0.3.4
ffmpeg-python==0.2.0
filelock==3.6.0
frozenlist==1.3.0
future==0.18.2
greenlet==1.1.2
grpclib==0.4.2
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.13.2
hyperframe==6.0.1
identify==2.4.11
idna==3.3
install==1.3.5
itsdangerous==2.1.0
jieba==0.42.1
Jinja2==3.0.3
lxml==4.8.0
MarkupSafe==2.1.0
multidict==6.0.2
nodeenv==1.6.0
pilk==0.0.2
ping3==3.0.2
platformdirs==2.5.1
pre-commit==2.17.0
priority==2.0.0
psycopg2==2.9.3
pydub @ file:///Users/hiro/Library/Caches/pypoetry/artifacts/2b/db/ff/deaaf55e81435b9e18466b557afca518988b6274385c3976f4e444b2fb/pydub-0.25.1-py2.py3-none-any.whl
pyee==9.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
quart==0.16.3
requests==2.27.1
semver==2.13.0
six==1.16.0
SQLAlchemy==1.4.31
stringcase==1.2.0
toml==0.10.2
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
virtualenv==20.13.2
wechaty==0.8.34
wechaty-grpc==0.20.19
wechaty-plugin-contrib==0.0.13
wechaty-puppet==0.4.0
wechaty-puppet-service==0.8.8
Werkzeug==2.0.3
wsproto==1.1.0
yarl==1.7.2
